requirements/dev.txt
-r base.txt
django-debug-toolbar
pytest>=8.0.0
pytest-django>=4.8.0
black>=24.1.1
flake8>=7.0.0
isort>=5.13.2
jsonschema>=4.17.3
requests>=2.31.0
